# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FootballLive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /spotlights:
    get:
      operationId: listSpotlights
      tags: [spotlights]
      summary: List all spotlights
      responses:
        "200":
          description: List of spotlights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spotlight"
    post:
      operationId: createSpotlight
      tags: [spotlights]
      summary: Create a spotlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotlightInput"
      responses:
        "201":
          description: Created spotlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spotlight"
  /spotlights/{id}:
    patch:
      operationId: updateSpotlight
      tags: [spotlights]
      summary: Update a spotlight
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotlightInput"
      responses:
        "200":
          description: Updated spotlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spotlight"
    delete:
      operationId: deleteSpotlight
      tags: [spotlights]
      summary: Delete a spotlight
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List all matches
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [live, scheduled, finished, all]
        - name: competition
          in: query
          schema:
            type: string
        - name: sport
          in: query
          schema:
            type: string
            enum: [football, futsal, all]
        - name: tournamentId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
...
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a single match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Not found
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    delete:
      operationId: deleteMatch
      tags: [matches]
      summary: Delete a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, shortName, logoUrl, country, sport]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        logoUrl:
          type: string
        country:
          type: string
        sport:
          type: string
          enum: [football, futsal]
...
    Match:
      type: object
      required: [id, homeTeam, awayTeam, homeScore, awayScore, status, minute, competition, kickoffAt, streamCount, sport, homeRedCards, awayRedCards, homeYellowCards, awayYellowCards]
      properties:
        id:
          type: integer
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        homeScore:
          type: integer
        awayScore:
          type: integer
        status:
          type: string
          enum: [live, scheduled, finished, postponed]
        minute:
          type: ["string", "null"]
        competition:
          type: string
        competitionLogo:
          type: ["string", "null"]
        kickoffAt:
          type: string
          format: date-time
        streamCount:
          type: integer
        featured:
          type: boolean
        sport:
          type: string
          enum: [football, futsal]
        tournamentId:
          type: ["integer", "null"]
        venue:
          type: ["string", "null"]
        matchGroup:
          type: ["string", "null"]
        homeRedCards:
          type: integer
        awayRedCards:
          type: integer
        homeYellowCards:
          type: integer
        awayYellowCards:
          type: integer
    MatchDetail:
      type: object
      required: [id, homeTeam, awayTeam, homeScore, awayScore, status, minute, competition, kickoffAt, streams, events, sport]
      properties:
        id:
          type: integer
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        homeScore:
          type: integer
        awayScore:
          type: integer
        status:
          type: string
          enum: [live, scheduled, finished, postponed]
        minute:
          type: ["string", "null"]
        competition:
          type: string
        competitionLogo:
          type: ["string", "null"]
        kickoffAt:
          type: string
          format: date-time
        streamCount:
          type: integer
        featured:
          type: boolean
        sport:
          type: string
          enum: [football, futsal]
        tournamentId:
          type: ["integer", "null"]
        venue:
          type: ["string", "null"]
        matchGroup:
          type: ["string", "null"]
        streams:
          type: array
          items:
            $ref: "#/components/schemas/Stream"
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEvent"
    MatchInput:
      type: object
      required: [competition, kickoffAt, sport]
      properties:
        homeTeamId:
          type: ["integer", "null"]
        awayTeamId:
          type: ["integer", "null"]
        homeScore:
          type: integer
        awayScore:
          type: integer
        status:
          type: string
          enum: [live, scheduled, finished, postponed]
        minute:
          type: string
        competition:
          type: string
        competitionLogo:
          type: string
        kickoffAt:
          type: string
          format: date-time
        featured:
          type: boolean
        sport:
          type: string
          enum: [football, futsal]
        tournamentId:
          type: integer
        venue:
          type: string
        matchGroup:
          type: string
    MatchUpdate:
      type: object
      properties:
        homeScore:
          type: integer
        awayScore:
          type: integer
        status:
          type: string
          enum: [live, scheduled, finished, postponed]
        minute:
          type: string
        featured:
          type: boolean
        venue:
          type: string
        kickoffAt:
          type: string
          format: date-time
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        tournamentId:
          type: ["integer", "null"]
        competition:
          type: string
        matchGroup:
          type: ["string", "null"]
    MatchEvent:
      type: object
      required: [id, matchId, type, minute, teamId, playerName]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        type:
          type: string
          enum: [goal, yellow_card, red_card, second_yellow_red, own_goal, penalty_awarded, penalty_goal, penalty_missed, ten_meter_goal, foul, substitution, mvp]
        minute:
          type: string
        teamId:
          type: integer
        playerName:
          type: string
        playerNumber:
          type: ["string", "null"]
        assistPlayerName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...
    Stream:
      type: object
      required: [id, matchId, label, url, quality, language]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        label:
          type: string
        url:
          type: string
        quality:
          type: string
          enum: [HD, SD, FHD]
        language:
          type: string
        embedCode:
          type: ["string", "null"]
...
    Spotlight:
      type: object
      required: [id, title, imageUrl, active, sortOrder, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    SpotlightInput:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        active:
          type: boolean
        sortOrder:
          type: integer
...