# lib/api-zod/src/generated/types/matchUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FootballLive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchUpdateStatus } from './matchUpdateStatus';

export interface MatchUpdate {
  homeScore?: number;
  awayScore?: number;
  status?: MatchUpdateStatus;
  minute?: string;
  featured?: boolean;
  venue?: string;
  kickoffAt?: Date;
  homeTeamId?: number;
  awayTeamId?: number;
  /** @nullable */
  tournamentId?: number | null;
  competition?: string;
  /** @nullable */
  matchGroup?: string | null;
}